sRNAtoolboxweb/requirements.txt
Django==1.11.2
pytz==2017.2
wheel==0.24.0
dajax==1.3
xlrd==1.0.0
pygal==2.3.1
djangorestframework==3.6.3
django-tables2==1.7.1
django-crispy-forms==1.6.1